further minor fixes
