+++ client/store/index.js
@@ -2,8 +2,8 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import game from './game'
+import user from './reducers/user'
+import game from './reducers/game'
 
 const reducer = combineReducers({user, game})
 const middleware = composeWithDevTools(
@@ -12,4 +12,4 @@ const middleware = composeWithDevTools(
 const store = createStore(reducer, middleware)
 
 export default store
-export * from './user'
+export * from './reducers/user'
