+++ client/store/reducers/game.js
@@ -1,0 +1,8 @@
+import player from './player'
+import opponent from './opponent'
+import data from './gameData'
+
+import {combineReducers} from 'redux'
+
+// export default oldReducer;
+export default combineReducers({player, opponent, data})
