+++ client/store/reducers/user.spec.js
@@ -6,7 +6,7 @@ import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
 import configureMockStore from 'redux-mock-store'
 import thunkMiddleware from 'redux-thunk'
-import history from '../history'
+import history from '../../history'
 
 const middlewares = [thunkMiddleware]
 const mockStore = configureMockStore(middlewares)
